agents/memory/requirements.txt
fastapi
psycopg2-binary
pgvector
neo4j